requirements.txt
<<<<<<< HEAD
=======
aiofiles==23.2.1
>>>>>>> 234d87dad4f0167267520124ba349a6d236955fe
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
<<<<<<< HEAD
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
auth0-python==4.7.2
backoff==2.2.1
bcrypt==4.2.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
crewai==0.98.0
cryptography==43.0.3
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.15
distro==1.9.0
docstring_parser==0.16
durationpy==0.9
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.115.6
filelock==3.17.0
flatbuffers==25.1.21
=======
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.9.0.post1
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
>>>>>>> 234d87dad4f0167267520124ba349a6d236955fe
frozenlist==1.5.0
fsspec==2024.12.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
<<<<<<< HEAD
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
=======
gradio==5.12.0
gradio_client==1.5.4
greenlet==3.1.1
>>>>>>> 234d87dad4f0167267520124ba349a6d236955fe
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
<<<<<<< HEAD
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
instructor==1.7.2
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json_repair==0.35.0
jsonpatch==1.33
jsonpickle==4.0.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
=======
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
>>>>>>> 234d87dad4f0167267520124ba349a6d236955fe
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.5
langsmith==0.2.10
litellm==1.57.4
markdown-it-py==3.0.0
<<<<<<< HEAD
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
=======
MarkupSafe==2.1.5
marshmallow==3.25.1
mdurl==0.1.2
>>>>>>> 234d87dad4f0167267520124ba349a6d236955fe
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.1
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.59.6
<<<<<<< HEAD
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.14
overrides==7.7.0
packaging==24.2
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.5
phonenumbers==8.13.52
pillow==11.1.0
posthog==3.9.0
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
pure_eval==0.2.3
=======
orjson==3.10.14
packaging==24.2
pandas==2.2.3
phonenumbers==8.13.52
pillow==11.1.0
propcache==0.2.1
>>>>>>> 234d87dad4f0167267520124ba349a6d236955fe
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
<<<<<<< HEAD
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
=======
PyMySQL==1.1.1
>>>>>>> 234d87dad4f0167267520124ba349a6d236955fe
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
<<<<<<< HEAD
pyvis==0.3.2
PyYAML==6.0.2
referencing==0.36.1
=======
PyYAML==6.0.2
>>>>>>> 234d87dad4f0167267520124ba349a6d236955fe
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
<<<<<<< HEAD
rpds-py==0.22.3
rsa==4.9
ruff==0.9.1
=======
rsa==4.9
ruff==0.9.1
safehttpx==0.1.6
>>>>>>> 234d87dad4f0167267520124ba349a6d236955fe
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
semantic-version==2.10.0
sentence-transformers==3.3.1
setuptools==75.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
<<<<<<< HEAD
stack-data==0.6.3
=======
sqlmodel==0.0.22
>>>>>>> 234d87dad4f0167267520124ba349a6d236955fe
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
<<<<<<< HEAD
tomli==2.2.1
tomli_w==1.2.0
tomlkit==0.13.2
torch==2.5.1
tqdm==4.67.1
traitlets==5.14.3
=======
tomlkit==0.13.2
torch==2.5.1
tqdm==4.67.1
>>>>>>> 234d87dad4f0167267520124ba349a6d236955fe
transformers==4.48.0
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
<<<<<<< HEAD
uv==0.5.22
=======
>>>>>>> 234d87dad4f0167267520124ba349a6d236955fe
uvicorn==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
wheel==0.44.0
<<<<<<< HEAD
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
=======
yarl==1.18.3
>>>>>>> 234d87dad4f0167267520124ba349a6d236955fe
